+++ x/fee/handler.go
@@ -16,12 +16,16 @@ func NewHandler(keeper *Keeper) sdkTypes.Handler {
 			return handleMsgAssignFeeToMsg(ctx, keeper, msg)
 		case MsgAssignFeeToAcc:
 			return handleMsgAssignFeeToAcc(ctx, keeper, msg)
+		case MsgAssignFeeToToken:
+			return handleMsgAssignFeeToToken(ctx, keeper, msg)
 		case MsgMultiplier:
 			return handleMsgMultiplier(ctx, keeper, msg)
 		case MsgTokenMultiplier:
 			return handleMsgTokenMultiplier(ctx, keeper, msg)
 		case MsgDeleteSysFeeSetting:
 			return handleMsgDeleteSysFeeSetting(ctx, keeper, msg)
+		case MsgDeleteAccFeeSetting:
+			return handleMsgDeleteAccFeeSetting(ctx, keeper, msg)
 		default:
 			errMsg := fmt.Sprintf("Unrecognized fee Msg type: %v", msg.Type())
 			return sdkTypes.ErrUnknownRequest(errMsg).Result()
@@ -41,6 +45,10 @@ func handleMsgAssignFeeToAcc(ctx sdkTypes.Context, keeper *Keeper, msg MsgAssign
 	return keeper.AssignFeeToAcc(ctx, msg)
 }
 
+func handleMsgAssignFeeToToken(ctx sdkTypes.Context, keeper *Keeper, msg MsgAssignFeeToToken) sdkTypes.Result {
+	return keeper.AssignFeeToToken(ctx, msg)
+}
+
 func handleMsgMultiplier(ctx sdkTypes.Context, keeper *Keeper, msg MsgMultiplier) sdkTypes.Result {
 	return keeper.CreateMultiplier(ctx, msg)
 }
@@ -51,4 +59,8 @@ func handleMsgTokenMultiplier(ctx sdkTypes.Context, keeper *Keeper, msg MsgToken
 
 func handleMsgDeleteSysFeeSetting(ctx sdkTypes.Context, keeper *Keeper, msg MsgDeleteSysFeeSetting) sdkTypes.Result {
 	return keeper.DeleteFeeSetting(ctx, msg)
+}
+
+func handleMsgDeleteAccFeeSetting(ctx sdkTypes.Context, keeper *Keeper, msg MsgDeleteAccFeeSetting) sdkTypes.Result {
+	return keeper.DeleteAccFeeSetting(ctx, msg)
 }